requirements.txt
Django==5.0.6
django-ninja==1.2.0
django-ninja-jwt[crypto]
pydantic[email]
django-cors-headers